Refactor : 로그인 핸들, 유효성 함수 분리
