+++ src/pages/login/checkToken.js
@@ -1,0 +1,9 @@
+export default function checkToken() {
+    const token = localStorage.token;
+
+    if (!token) {
+        return false;
+    }
+
+    return true;
+}
