+++ src/pages/login/getMyInfo.js
@@ -1,0 +1,20 @@
+export default async function getMyInfo() {
+    const requestPath = '/user/myinfo';
+    const requestUrl = `${URL}${requestPath}`;
+
+    const token = await localStorage.token;
+    const bearerToken = `Bearer ${token}`;
+
+    const response = await fetch(requestUrl, {
+        method: 'GET',
+        headers: {
+            Authorization: bearerToken,
+            'Content-type': 'application/json',
+        },
+        body: JSON.stringify(),
+    });
+
+    const json = await response.json();
+
+    return json.user;
+}
